aideck/backend/modules/security/jwt_handler.py
"""
JWT handler for AIDECK
"""
def create_jwt_token(data: dict):
    from datetime import datetime, timedelta
    from jose import jwt
    from config import settings
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
    return encoded_jwt

def verify_jwt_token(token: str):
    from jose import jwt, JWTError
    from config import settings
    try:
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
        return payload
    except JWTError:
        return None
