aideck/backend/requirements.txt
psycopg2-binary
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
databases[aiosqlite]
python-jose[cryptography]
passlib[bcrypt]
pydantic
alembic
httpx
python-dotenv
celery
redis
openai
fastapi-limiter
pydantic-settings
asyncpg
python-multipart
